requirements.txt
langchain-openai>=0.0.2
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0 